refactor(agent): 重构Agent应用启动及自动装配配置
- 从AgentApplication中移除启动runner的内部类，迁移至AgentRunnerConfig配置类
- 在AgentAutoConfiguration中移除@ComponentScan，改用@Bean方式注入各采集器与服务
- 移除多个类上的@Component注解，统一通过配置类注册Bean，简化Spring管理
- 修正InstanceTaskService拉取任务URL构造，补全参数等号
- 添加对SocketTimeoutException的捕获，避免长轮询超时异常日志泛滥
- 修改server端实例及任务相关模型，将端口字段由i16变更为i32，提升兼容性
- 优化server端实例任务查询及分页逻辑，及参数校验与日志打印格式
- 微调部分Rust代码格式与错误处理，提高代码规范性和健壮性

+++ apps/server/src/entities/instances.rs
@@ -13,7 +13,7 @@ pub struct Model {
     pub application_id: String,
     pub mac_address: Option<String>,
     pub public_ip: Option<String>,
-    pub port: Option<i16>,
+    pub port: Option<i32>,
     pub program_path: Option<String>,
     pub profiles: Option<String>,
     pub os_type: Option<String>,
